python_requirements.txt
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
anyio==3.3.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.7
backcall==0.2.0
blessings==1.7
blis==0.7.4
Brotli==1.0.9
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorful==0.5.4
contextlib2==21.6.0
contextvars==2.4
cryptography==3.4.6
cycler==0.10.0
cymem==2.0.5
Cython==0.29.22
dataclasses==0.8
decorator==4.4.2
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.70.0
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.9.0
Flask-Cors==3.0.10
Flask-WTF==0.14.3
gensim==3.8.3
gevent==21.1.2
google-api-core==1.26.3
google-auth==1.28.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
greenlet==1.0.0
grpcio==1.36.1
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
httptools==0.2.0
idna==2.10
idna-ssl==1.1.0
immutables==0.15
importlib-metadata==3.7.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
kiwisolver==1.3.1
llvmlite==0.34.0
MarkupSafe==1.1.1
matplotlib==3.3.4
msgpack==1.0.2
multidict==5.1.0
murmurhash==1.0.5
nltk==3.5
nodeenv==1.6.0
numba==0.51.2
numpy==1.19.5
nvidia-ml-py3==7.352.0
opencensus==0.7.12
opencensus-context==0.1.2
openTSNE==0.5.1
orjson==3.6.1
packaging==20.9
pandas==1.1.5
parso==0.8.1
pathy==0.4.0
pdfminer.six==20201018
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pkg_resources==0.0.0
preshed==3.0.5
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.6
psutil==5.8.0
ptyprocess==0.7.0
py-spy==0.3.5
pyarrow==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.7.4
Pygments==2.8.0
pymongo==3.11.3
pynndescent==0.5.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2020.11.13
requests==2.25.1
rsa==4.7.2
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.5.4
sentence-transformers==0.4.1.2
sentencepiece==0.1.95
six==1.15.0
smart-open==3.0.0
sniffio==1.2.0
sortedcontainers==2.3.0
srsly==2.4.0
starlette==0.16.0
thinc==8.0.1
threadpoolctl==2.1.0
tokenizers==0.10.1
toml==0.10.2
torch==1.7.1
tqdm==4.57.0
traitlets==4.3.3
transformers==4.3.2
typer==0.3.2
typing-extensions==3.7.4.3
ujson==4.0.2
umap-learn==0.5.1
urllib3==1.26.3
uvicorn==0.15.0
uvloop==0.14.0
wasabi==0.8.2
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
Werkzeug==1.0.1
WTForms==2.3.3
yarl==1.6.3
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.3.0
